scripts/requirements.txt
# Python dependencies for tax rate harvester
groq>=0.11.0
requests>=2.31.0
pyyaml>=6.0.1
